Improve ResourceLinkData (of RichText) link resolution
Add case .entryDecodable to Link
Resolve link in ResourceLinkData.target directly
Eliminate ResourceLinkData.resolvedResource
Adapt test cases accordingly
Add a new test case that checks if multi-level linking in RichText
  content works (testDecodingNestedEntriesInlinedInRichText)

Discussion:
In order to get rid of ResourceLinkData.resolvedResource, which can be
considered a hack (since ResourceLinkData.target should be capable
of holding all the required information), a new case is added to
Link, .entryDecodable. In this new case, user-customized types
conforming to EntryDecodable can be stored after they have been
resolved. So, if a RichText document links or inlines to an
EntryDecodable, ResourceLinkData.target can now hold the reference to
the actual EntryDecodable object.
Link resolution in ResourceLinkData is adapted accordingly.
This is a breaking API change, but it arguably fixes a major
consistency problem in link resolution.